added custom exception and global handler, util class and services impl

+++ src/main/java/com/smallworld/service/IssueService.java
@@ -4,6 +4,6 @@ import java.util.List;
 import java.util.Set;
 
 public interface IssueService {
-    Set<Integer> getUnsolvedIssueIds();
+    Set<Long> getUnsolvedIssueIds();
     List<String> getAllSolvedIssueMessages();
 }
